feat(utils): separate "avatar"
