+++ util/github.ts
@@ -1,0 +1,2 @@
+export const githubProfileAvatarUrl = (username: string): string =>
+  `https://avatars.githubusercontent.com/${username}`;
